Enigma_plus
First and foremost this is a simple enigma simulation program written in python. It started as a fun project to see if I could emulate the enigma machine. A few additional features have been added. The option to have as many rotors as desired with the ability to choose the character sets for those rotors. Additionally, there is a rotor generator to generate rotors based based on the user selector seed.